requirements.txt
requests~=2.27.1
PyJWT~=2.3.0
PyMySQL~=1.0.2
peewee~=3.14.10
peewee-async~=0.7.2
bcrypt~=3.2.0
aiofiles~=0.8.0
aiomysql~=0.1.0
redis~=4.3.4
tornado~=6.2
WTForms~=2.3.3
WTForms-JSON~=0.3.5
wtforms-tornado~=0.0.2
celery~=5.2.7
flower~=1.2.0
supervisor~=4.2.4
loguru~=0.6.0
PyYAML~=6.0